+++ src/Application/Components/Contact/CanParticipantSeeContact.php
@@ -23,7 +23,11 @@ class CanParticipantSeeContact
             new GetContactListUsersViewQuery($seerParticipant->getEvent(), $seerParticipant)
         );
 
-        $metUsers = array_merge($usersView->inContactsUsers, $usersView->requestsUsers);
+        $metUsers = array_merge(
+            $usersView->inContactsUsers,
+            $usersView->requestsUsers,
+            $usersView->metInMeetingUsers
+        );
 
         return \in_array($seenUser, $metUsers, true);
     }