+++ src/Application/Components/Navigation/Route.php
@@ -43,6 +43,8 @@ final class Route
     public const MEETING_REQUEST_LIST = 'event_meeting_list_request';
     public const MEETING_REQUEST_MERGED_LIST = 'event_meeting_request_merged_list';
 
+    public const MEETING_EVALUATION = 'event_meeting_evaluation';
+
     public const PACKAGE = [
         'event_package_step',
         'event_package_add_participant',
@@ -74,6 +76,7 @@ final class Route
     ];
 
     public const AGENDA_DEFAULT = 'event_agenda';
+    public const AGENDA_PARTICIPANT = 'event_agenda_participant';
     public const AGENDA_CONFIRMATION = 'event_user_event_token_confirm_agenda';
 
     public const PROGRAM = 'happening_program';