+++ src/Application/Query/Happening/Webinar/GetWebinarViewQueryHandler.php
@@ -3,6 +3,7 @@
 namespace Proximum\Vimeet\Application\Query\Happening\Webinar;
 
 use Proximum\Vimeet\Application\Adapter\VideoConferenceAdapterInterface;
+use Proximum\Vimeet\Application\View\Happening\WebinarSpeakerView;
 use Proximum\Vimeet\Application\View\Happening\WebinarView;
 use Proximum\Vimeet\Domain\Time\TimeRangeView;
 
@@ -14,8 +15,10 @@ class GetWebinarViewQueryHandler
     /** @var \DateTimeInterface */
     private $dateTime;
 
-    public function __construct(VideoConferenceAdapterInterface $videoConferenceAdapter, \DateTimeInterface $dateTime)
-    {
+    public function __construct(
+        VideoConferenceAdapterInterface $videoConferenceAdapter,
+        \DateTimeInterface $dateTime
+    ) {
         $this->videoConferenceAdapter = $videoConferenceAdapter;
         $this->dateTime = $dateTime;
     }
@@ -41,13 +44,27 @@ class GetWebinarViewQueryHandler
         $sessionId = $session->getSessionId();
         $timeRemainingInSeconds = max(0, $happening->getEnd()->getTimestamp() - $this->dateTime->getTimestamp());
 
+        $speakers = [];
+
+        foreach ($happening->getSpeakers() as $speaker) {
+            $speakers[] = new WebinarSpeakerView(
+                $speaker->getUser()->getId(),
+                $speaker->getFirstname(),
+                $speaker->getLastname(),
+                $speaker->getPosition($query->getLocale()),
+                $speaker->getOrganization()
+            );
+        }
+
         return new WebinarView(
             $happening->getId(),
+            $query->getUser()->getId(),
             $happening->getTitle($query->getLocale()),
             $token,
             $sessionId,
             $this->videoConferenceAdapter->getApiKey(),
             $isSpeaker,
+            $speakers,
             new TimeRangeView($happening->getBegin(), $happening->getEnd()),
             $this->dateTime,
             $timeRemainingInSeconds,